--- src/utils/cw_misc.hpp
@@ -1,5 +1,0 @@
-// ==================================================================
-// Author: Ashley Zhang (ayz27@cornell.edu)
-// Date:   2/15/2025
-// Description: cw misc functions and objects not deserving of separate file
-// ==================================================================
